[PWA] Possible crash fix in icon masking code on Mac
This is a speculative fix for a hard to repro crash in icon masking code
on Mac. See crash/6865c2025e19612a as an example.

Calling ToSkBitmap() causes this CHECK to fail [1] if the gfx::Image
created is empty. This works around that by using AsBitmap(), which
instead of crashing returns an empty SkBitmap if the gfx::Image
representation for it does not exist. On top of this, to align with
behavior on Windows and Linux, if the masked bitmap is empty, the input
bitmap is passed directly to the final result callback, as a stopgap.

[1] https://source.chromium.org/chromium/chromium/src/+/main:ui/gfx/image/image.cc;l=452;bpv=1;bpt=1?q=gfx::Image

Fixed: 457966490
Change-Id: I83a18598c0ab60a6eea430438d4a45f1171c3c5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7125239
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540755}