Screenshot read back: Setup SkBitmap in background thread
Allocating and advising a bitmap to write to use for the read back can
be quite expensive. Even though we only issue the read back during
quiescence, if the situation suddenly changes, the UI thread could be
busy for some time. The rest of the operations (actually issuing the
read back and receiving the written bitmap) are much shorter.

This should minimize any regressions on INP.

Bug: 438496406
Change-Id: Iff8749694ee34971444ab62cfabdcfe86a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7121920
Reviewed-by: William Liu <liuwilliam@chromium.org>
Commit-Queue: Aldo Culquicondor <acondor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540786}