Update Windows GPU collection function with extra attribute for NPUs
In testing out some of the new functionality in Chromium for using WebNN
with the Windows ML backend, we discovered that on an AMD device with an
NPU, the NPU was not being detected by the code in
`CollectNPUInformation`.

DXCore has an additional hardware attribute GUID for NPUs, which is able
to resolve the NPU on this machine in local testing. Generally speaking,
newer NPUs will have this attribute set, so this change introduces the
attribute and further refactors the code to iterate / aggregate NPUs
over all three attributes.

How verified: I ran a private build on the device where the NPU was
previously not being found and checked the `chrome://gpu` settings and
confirmed that the output contains an NPU0 entry with metadata for the
device. I also validated my LunarLake Intel PC, my Qualcomm PC, and tried running it on my desktop workstation which has the TRT RTX EP installed on it (that EP does not pair with NPUs, though).

Change-Id: I3bc0e7e406d781ded1753d7858b7980385f381b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7075603
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Commit-Queue: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: Hu, Ningxin <ningxin.hu@intel.com>
Cr-Commit-Position: refs/heads/main@{#1540787}