[owners] Remove fjacky@chromium.org from components/permissions/PERMISSIONS_OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/fjacky@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I28293e84fff45ac8902738738fedfeb2f595b7b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7118521
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Ravjit Uppal <ravjit@chromium.org>
Reviewed-by: Ravjit Uppal <ravjit@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540794}