[Desktop Android] Switch extension installed dialog to use DialogModel
This is the first part of porting the extension installed bubble to
Desktop Android.

Screenshots:

- w/o sign-in promo:
https://drive.google.com/file/d/1Ey2_2Y4jKO1frMU2vS0S1Ta6TgORbB9O/view

- w/ sign-in promo (theme colors):
https://drive.google.com/file/d/1Lp08YKb2t5-VGwB3ftmRNVwys2iiizTl/view

- w/ sign-in promo (dark mode):
https://drive.google.com/file/d/14jEF-3jc2q2QJhXZTJ9kAZPWWkTVsw3m/view

Tested that sign-in link works:
https://drive.google.com/file/d/1lmz7W5TNh7db-7tiVXBs64OcmIu9Kvl5/view

Width of dialog is slightly shorter compared to prod
(https://drive.google.com/file/d/1VhYcC0kunJr47k8ECcDrMVqtpfu2zflq/view)
but this is okay as this is the default `DialogModel` width, and dialogs
should all be aligned to this standard.

Bug: 450296898
Skip-Translation-Screenshots-Check: True
Change-Id: Ib94040d9b453986d266a41420dd19778b507c467
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7124268
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Reviewed-by: Tim <tjudkins@chromium.org>
Commit-Queue: Eva Su <evasu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540821}