[Extensions] Refactor EventEmitter to expose HasListeners()
Previously, EventEmitter exposed a GetNumListeners() method that
returned the number of listeners for an event. This was used by
other classes to check if any listeners were present. A TODO comment
suggested that this should be a test-only method, and a boolean
HasListeners() method should be exposed for production code.

This change implements that suggestion. GetNumListeners() is renamed to
GetNumListenersForTesting() and its usage is restricted to unit tests.
A new public HasListeners() method is exposed on EventEmitter to
provide a cleaner API for checking the existence of listeners. Call
sites in APIEventHandler and various unit tests have been updated to
use the new, more appropriate methods.

Bug: N/A
Change-Id: Ied1d5e3517a96f5646946c3eb7a0ce232f91abac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7047100
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Auto-Submit: Justin Lulejian <jlulejian@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540826}