Fix crash when FrameData::input_state is null
When the XRSession is not fully visible we hide the input sources from
the page. While this member is `optional`, remaining parts of the
`GetNextFrameData` call did not treat this as such. This adds plumbing
to the `GetNextFrameData` call to properly treat this data as possibly
null.

Fixed: 457945081
Change-Id: I77eb028b0be63d38c81d0c7a9dae3ec83ff3ed36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7125301
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540857}