Remove some dependencies from the custom_handlers component
The extensions::ProtocolHandlerManager use the ProtocolHandlerRegistry
to register / unregister protocol handlers associated to the enabled
extensions during the ExtensionSystem initialization. There was a DCHECK
to ensure the ProtocolHandlerRegistry instance is available, or can be
created by the corresponding factory, in the profile. However, some
tests use dummy profiles that lacks the required testing factory.

We had introduced a SimpleProtocolHandlerRegistryFactory as testing
factory in the unit tests that may require it, in order to avoid
violating the mentioned assert, even though the tests themselves didn't
require the use of the ProtocolHandlerRegistry instance at all.

I think it's better to relax the assert, with a comment mentioning that
the registry instance can be null under testing environments, instead of
adding a dependency of the custom_handlers component in so many places.

Bug: 40482153

Change-Id: I76286524fef59a5b64727bf1aaf38df61fdefc08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7116098
Commit-Queue: Javier Fernandez <jfernandez@igalia.com>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540858}