Add skipTest() to WebGL perf tests
I didn't realize this utility came from webgpu-perf-utils.js and not
from runner.js. It didn't exist, which caused an exception causing a
test timeout (same as what happened before when `gl` was null).
Copy it to WebGL so we can use it here too.

Passes with:
  ./tools/perf/run_benchmark blink_perf.webgl --browser canary
Skips with:
  ./tools/perf/run_benchmark blink_perf.webgl --browser canary --extra-browser-args=--disable-gpu

Bug: 456419577
Change-Id: Ia105b6343481fd031b1d0b91e68a71009b2b2c16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7124800
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540908}