[HIC] use CSS-to-canvas-grid scale factor for drawing
Prior to this patch, if a destination size was not provided to
drawElementImage, its size would depend on deviceScaleFactor. With this
change, if no destination size is provided the element will be drawn at
the same scale as it would appear if painted outside the canvas.

Spec discussion at https://github.com/WICG/html-in-canvas/issues/30

Bug: 456097302
Change-Id: Ic41efcd8fdc7be9e0bcb1ea7da1c93c3b6aa3d71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7125102
Commit-Queue: Stefan Zager <szager@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540931}