ash: Prevent brief flickering due to initial drag window positioning
In ChromeOS tablet mode (non-WebUI tab strip), when dragging a tab out
of a snapped browser (split view), the drag window is briefly maximized
before being resized, which causes a brief visual distraction. Prevent
this by setting the state of the drag window to snapped.

The same issue exists for floated browsers. However, we can't set the
drag window to floated, as that would violate the invariant that there's
only one floated window. We'll address this separately.

Bug: b:421465978
Change-Id: Iba1fb3ba27fbc0d11fb496511c3c4329b829560e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7099681
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Georg Neis <neis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540956}