[Contextual Tasks] Pass along whether a navigation is a link
This patch adds a parameter to the navigation handling method in the
ContextualTasksUiService to identify whether the navigation originated
from the page or from some top chrome UI (e.g. back, forward, omnibox,
etc.). This prevents us trying to handle the navigation if the user
hits the back button (which would also open the side panel if it was
closed). A test has been added for this case.

Bug: 445841267
Change-Id: Id3b856cb89d451b9b1ed5732235f445e8bfb39fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7125167
Reviewed-by: Duncan Mercer <mercerd@google.com>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1540959}