Implement history event serialization
Allow event objects to be serialized to base::Value::Dict. Conversion to
JSON and serialization to storage will be introduced in a follow-up CL.

While implementing this CL a difference between the documented format
for durations and the tools available in //base was discovered. The
documented format is updated to support the code in this case.

Bug: 456497715
Change-Id: I71b6c350ddc3a7e25ef597e8207202c3aee715a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7125380
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Commit-Queue: Noah Rose Ledesma <noahrose@google.com>
Cr-Commit-Position: refs/heads/main@{#1541030}