Rebind high rank processes while the app is visible but not focused
The StrictHighRankProcessLRU feature avoids rebinding high-rank renderer
processes when Chrome is in the foreground. This reduces binder IPCs, as
the process priority is determined by the binding flag (STRONG vs.
VISIBLE) rather than the LRU order.

However, this assumption is not always correct. When the Chrome
application is visible but not focused (e.g., when a share dialog is
open), the process's oom_score_adj is capped at the VISIBLE level. This
removes the priority distinction between STRONG and VISIBLE bindings,
and the LRU order is used to determine which process to kill. This can
lead to an increase in "Visible renderer foreground oom" crashes, as a
visible main frame can be killed before a visible subframe or even other
applications

This change fixes the issue by rebinding high-rank processes when the
application is visible but not focused. This ensures that high-rank
processes are kept at the top of the LRU list, preventing them from
being killed.

To detect if the application is focused, this CL uses
onWindowFocusChanged as a proxy for onTopResumedActivityChanged. This is
a short-term solution, as onWindowFocusChanged can be triggered by
non-activity windows like dialogs and popups. A long-term solution would
be to use onTopResumedActivityChanged directly, but this is not easily
implementable in the current architecture.

Bug: 431653029
Change-Id: Iabcfc87099120ab237d0fc389bf2c143436294ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7103621
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Shin Kawamura <kawasin@google.com>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541033}