ChromeAndroidTaskTracker: use MultiInstanceManager to create CTA Intent
Previously the Intent to launch a new instance of ChromeTabbedActivity
(CTA) was hard-coded in ChromeAndroidTaskTrackerImpl. The proper and
standard way is to get the Intent from MultiInstanceManager, as
MultiInstanceManager is the source of truth and handles all the
complexities involved in configuring the Intent (flags & Extras).

As MultiInstanceManager is associated with CTA, it's part of
ChromeAndroidTask's ActivityScopedObjects. This means to create a new
instance of CTA, an existing CTA must exist. In this CL, when creating
the Intent, if we can't find a ChromeAndroidTask that has
MultiInstanceManager, we'll return null. This behavior is expected as
we never launch a new instance of CTA without an existing CTA, except
the initial CTA launched during app start-up.

As a result of using MultiInstanceManager, CTA instance limit must be
disabled in tests run on an emulator:
* create_browser_window_android_browsertest.cc
* ChromeAndroidTaskIntegrationTest.java

Change-Id: I9088988453fb71a990032c672b0bcdec0de39b9b
Bug: 453777179
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7070228
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Linyu He <linyuh@google.com>
Reviewed-by: Aishwarya Rajesh <aishwaryarj@google.com>
Cr-Commit-Position: refs/heads/main@{#1541043}