Fix a bug in the permission element grd generating script.
There is a crash caused by a key look-up in the dictionary. This error
shouldn't happen. This CL doesn't address the issue, but it fixes the
build for now, by gracefully handling the error. The script breaks for
Strings with Placeholders in it. It will be solved in a follow-up CL.

Bug: 457422566
Change-Id: I323fbea9a472634686d852dc72ee4af69b54416e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7122314
Auto-Submit: Ravjit Uppal <ravjit@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541052}