ServiceWorker: Do not set a synthetic response flag in Incognito dry run
In the dry-run mode of a synthetic response experiment, the
`is_synthetic_response_used_` flag was being set to imitate the behavior
of the synthetic response in the renderer.

In Incognito mode, the browser does not signal to the server that it
supports synthetic responses. Setting this flag caused the renderer to
incorrectly expect Content Security Policies (CSPs) to be delivered via a
<meta> tag instead of HTTP headers.

This change corrects the behavior by not setting the flag when the browser
is in an off-the-record (Incognito) context.

Bug: 456965135
Change-Id: I23248d5149c0530ca6dfb800a5432b392c2e9785
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7122960
Auto-Submit: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541058}