[Cast] Fix GpuFactoryContextLost threading issue
This patch fixes an issue where the GPU factory context becomes lost
on the GPU thread, and then causes a sequence checking error tearing
down the audio input device (which is expecting to be torn down on the
main thread).

To make these types of threading issues easier to fix in the mirroring
component, a sequence checker is added to the OpenscreenSessionHost.

Bug: 457842368
Change-Id: Ia9b5fca7ca4d3f5b03128fa1f42cb950afa9c27f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7126219
Reviewed-by: Muyao Xu <muyaoxu@google.com>
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541060}