Introduce WebContentsBasedCanceller for dealing with navigation races.
This class captures the details of avoiding races with the dialog-owning
RenderFrameHosts's active state and the containing WebContents'
visibility.

Apply this to FileSystemChooser. This fixes the bug whereby the chooser
remains on-screen even after the document enters BFCache.

This also introduces an `IsActive` check in `ShowFilePickerOnUIThread`
to avoid triggering the `DCHECK(outermost_rfh->IsInPrimaryMainFrame())`.
This is redundant since the AutoCancel will also check this but I don't
think we should refactor to avoid this redundancy as it would require
creating the AutoCancel very early and far from where it is used.

This includes a hack of navigating away at the end to prevent the new
test, DoNotDismissOnSubframeNavigation, from triggering leak detection.
https://crrev.com/c/7122881 (which does nothing but add that test)
demonstrates that this is a pre-existing condition. I believe the leak
is not real and is just due to the fast-shutdown not triggering
destruction of the main frame without changing state. The follow-on
https://crrev.com/c/7106498 reverts the hack.

Bug: 449800514
Fixed: 443283015
Change-Id: Id2c05c10e9fcd0d163252af88330b882464d2b78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6986448
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Fergal Daly <fergal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541062}