[iOS] Use a real navigation bar for Safari Import
This CL updates how the navigation bar is done in Safari Import
ViewController.
Instead of relying on the navigation bar from the ConfirmationAlertVC,
this CL puts the ViewController in a NavigationController.

Note: there is an issue with how the height is computed, which makes it
too big in the old version and too small in the new one.

Before: https://screenshot.googleplex.com/9zF8g9QFgwFGegf
After: https://screenshot.googleplex.com/3cfZ9hTLCqno5o4

Bug: 444648926
Change-Id: Ifdfdaf83f5825b00c05cad4391fee97f3d26a907
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7119801
Reviewed-by: Ginny Huang <ginnyhuang@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Ewann Pellé <ewannpv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541104}